sample: use existing plunge-gradle-plungin version from maven
Can't use version that hasn't been published yet.

This is further complicated by current 1.0.1 version on MavenCentral [1]
has a dependency on 1.0.0 of `nz.co.trademe.plunge:plunge-parsing` which
only existed in jcenter.

Once we've bootstrapped a new version of that into MavenCentral then we
can get rid of the jcenter repo.

[1] https://search.maven.org/artifact/nz.co.trademe.plunge/plunge-gradle-plugin/1.0.1/jar